backend/vulscan/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
selenium==4.7.2
fpdf==1.7.2